requirements.txt
attrs==15.2.0
backports.ssl-match-hostname==3.5.0.1
cffi==1.5.2
click==6.6
cryptography==1.3.2
docker-py==1.8.1
idna==2.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==0.15.1
queuelib==1.4.2
requests==2.10.0
Scrapy==1.0.5
service-identity==16.0.0
six==1.10.0
Twisted==15.5.0
w3lib==1.13.0
websocket-client==0.37.0
zope.interface==4.1.3
